src/gfx: Create the concept of material
Instead of a single shader and texture per cg_model, create cg_material
that holds the shader, texture, and material values (taken from the OBJ
file).

cg_model holds multiple cg_material and has a table that links the
meshes to the available materials.

Each mesh is drawn with it's related material.